requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
feedparser>=6.0.0
pyyaml>=6.0
httpx>=0.27.0
discord.py>=2.3.0
flask>=3.0
schedule>=1.2.0
